dotted_path_to_list moves into commandline module
