+++ tests/test_document.py
@@ -1,7 +1,6 @@
 # coding: utf-8
 from coreapi import action, required
 from coreapi import Array, Document, Object, Link, Error
-from coreapi.utils import dotted_path_to_list
 import pytest
 
 
@@ -430,7 +429,7 @@ def test_link_field_items_must_be_valid():
 
 # Invalid calls to '.action()' should error.
 
-def test_keys_should_be_a_list_or_dotted_string(doc):
+def test_keys_should_be_a_list_or_string(doc):
     with pytest.raises(TypeError):
         action(doc, True)
 
@@ -448,26 +447,6 @@ def test_keys_should_be_valid_indexes(doc):
 def test_keys_should_access_a_link(doc):
     with pytest.raises(ValueError):
         action(doc, 'dict')
-
-
-# Test dotted path notation maps to list of keys correctly.
-
-def test_dotted_path_notation():
-    doc = Document({'rows': [Document({'edit': Link()})]})
-    keys = dotted_path_to_list(doc, 'rows.0.edit')
-    assert keys == ['rows', 0, 'edit']
-
-
-def test_dotted_path_notation_with_invalid_array_lookup():
-    doc = Document({'rows': [Document({'edit': Link()})]})
-    keys = dotted_path_to_list(doc, 'rows.zero.edit')
-    assert keys == ['rows', 'zero', 'edit']
-
-
-def test_dotted_path_notation_with_invalid_key():
-    doc = Document({'rows': [Document({'edit': Link()})]})
-    keys = dotted_path_to_list(doc, 'dummy.0.edit')
-    assert keys == ['dummy', '0', 'edit']
 
 
 # Documents and Objects have `.data` and `.links` attributes
